Address PR review comments
- Extract Charge insertion into its own optimization pass
  (jit/optimization/insert_charge.rs) following the pass-based model in
  jit/mod.rs. translate.rs no longer mixes gas analysis with bytecode
  translation; it just translates and the new pass rewrites blocks.
- Add a thorough module-level doc to insert_charge.rs explaining what
  charges are subsumed and how.
- Move the Charge variant to the end of both Bytecode enums (execution
  and optimization) and add an extensive doc comment listing every opcode
  whose gas is subsumed and every opcode that still charges per-instruction.
- Move charge cost-computation tests from the inline #[cfg(test)] module
  in optimization/translate.rs to a dedicated unit_tests/insert_charge_tests.rs.
- Protocol-gate charge_block in SuiGasMeter on use_charge_batching
  (gas_model_version > 11) so the Charge instruction is a no-op for gas
  metering on older protocol versions where charge_simple_instr still
  charges per-instruction.
